+++ coremltools/converters/mil/testing_utils.py
@@ -143,7 +143,7 @@ def random_gen(
     If allow_duplicate is set to false, it is guaranteed that value generated are all different.
     Default data type is np.float32.
     """
-    elem = np.prod(shape).astype(np.int)
+    elem = np.prod(shape).astype(np.int32)
     ret = []
     for _ in range(elem):
         while True:
